% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSModel_estimation.R
\name{create_SSMmodel}
\alias{create_SSMmodel}
\title{Create an SSMmodel object and auxiliary information for a call to optim}
\usage{
create_SSMmodel(ts_data, regression_variables, regression_data,
  trend_degree, seasonal_period, seasonal_type = "common", cycle_num,
  cycle_base_period, cycle_type = "common", return_optim_args = FALSE)
}
\arguments{
\item{ts_data}{matrix of time series values, one column per time series and
one row per time point.  NA's are allowed.}

\item{regression_variables}{character vector of variables to use in a
regression component of the model.  If missing, no regression component is
included.}

\item{regression_data}{data frame with variables for regression}

\item{trend_degree}{degree of trend component; if missing, no trend
component is included.}

\item{seasonal_period}{period of seasonality term in mdoel; if missing, no
seasonal component is included.}

\item{seasonal_type}{whether seasonal states are shared across time series
("common") or distinct for each time series ("distinct")}

\item{cycle_num}{number of cycle components to include.  The \code{k}'th
cycle component will have period \code{cycle_base_period / k}.}

\item{cycle_base_period}{base period for cycle components.}

\item{cycle_type}{whether cycle states are shared across time series
("common") or distinct for each time series ("distinct")}

\item{return_optim_args}{boolean; if TRUE, also return functions and initial
parameter values that could be used in a call to optim for parameter
estimation.}
}
\value{
named list.  If \code{return_optim_args} is FALSE, contains only
  \code{model}, an SSMmodel object.  If \code{return_optim_args} is TRUE,
  also contains: \code{update_pars}, a function to update the \code{model}
  parameters; \code{calc_loglik}, a function to calculate the \code{model}
  log likelihood; and \code{init_pars}, a vector of initial parameter values.
}
\description{
Create an SSMmodel object and auxiliary information for a call to optim
}
