% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSModel_simulate_forecast.R
\name{simulate_alpha_t}
\alias{simulate_alpha_t}
\title{generate a single simulated value for alpha_t based on a mode fit and state vector alpha_{t-1}}
\usage{
simulate_alpha_t(model_fit, alpha_tm1)
}
\arguments{
\item{model_fit}{a SSMmodel fit object}

\item{alpha_tm1}{a vector (or matrix with one column) with states at time t - 1}
}
\value{
simulated values for the state at time t
}
\description{
generate a single simulated value for alpha_t based on a mode fit and state vector alpha_{t-1}
}
