% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSModel_simulate_forecast.R
\name{simulate_incidence}
\alias{simulate_incidence}
\title{simulate_incidence function}
\usage{
simulate_incidence(model_fit, backcast_steps, forecast_steps, n_sim,
  new_regression_data)
}
\arguments{
\item{backcast_steps}{integer number of steps to "backcast"; simulations of
incidence will be generated for the last backcast_steps time points}

\item{forecast_steps}{integer number of steps to forecast}

\item{n_sim}{integer number of paths to simulate}

\item{new_regression_data}{a data frame of regression variables only.}
}
\value{
A horizon x num_ts x n_sim array of sampled incidence.
}
\description{
simulate_incidence function
}
