% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSModel_simulate_forecast.R
\name{simulate_y_t}
\alias{simulate_y_t}
\title{generate a single simulated value for y_t based on a mode fit and state vector alpha_t}
\usage{
simulate_y_t(model_fit, alpha_t)
}
\arguments{
\item{model_fit}{a SSMmodel fit object}

\item{alpha_t}{a vector (or matrix with one column) with states at time t}
}
\value{
simulated values for the observed quantities y_t
}
\description{
generate a single simulated value for y_t based on a mode fit and state vector alpha_t
}
